+++ R/taxatree_plotkey.R
@@ -25,7 +25,64 @@
 #' @export
 #'
 #' @examples
-#' # see taxatree_plots() examples
+#' library(ggplot2)
+#' # Normally you make a key to accompany taxatree_plots showing stats results
+#' # So see ?taxatree_plots examples too!
+#' #
+#' # You can also use the taxatree key visualization just to help understand your
+#' # tax_table contents and hierarchical structure
+#' shao19 %>%
+#'   ps_filter(family_id %in% 1:5) %>%
+#'   tax_filter(min_prevalence = 7) %>%
+#'   tax_fix() %>%
+#'   tax_agg("genus") %>%
+#'   tax_prepend_ranks() %>%
+#'   taxatree_plotkey(rank %in% c("phylum", "genus"))
+#'
+#'
+#' # # Let's look at some of the most prevalent Actinobacteria
+#' actinoOnlyPhylo <- shao19 %>%
+#'   tax_select("Actinobacteria", ranks_searched = "phylum") %>%
+#'   tax_filter(min_prevalence = 100)
+#'
+#' actinoOnlyPhylo %>%
+#'   tax_mutate(order = NULL) %>%
+#'   tax_sort(by = "prev", at = "species", tree_warn = FALSE) %>%
+#'   taxatree_plotkey(
+#'     circular = FALSE, rank != "genus",
+#'     .draw_label = FALSE, # suppress drawing now so we can make custom labels after
+#'     size_guide = "legend", colour = "skyblue", edge_alpha = 0.2
+#'   ) %>%
+#'   taxatree_plot_labels(
+#'     circular = FALSE, # be sure you match the plotkey layout
+#'     fun = geom_text, fontface = "bold.italic", size = 2.5, hjust = 0,
+#'     nudge_y = 0.1 # nudge y if you want to nudge x (due to coord_flip!)
+#'   ) +
+#'   coord_flip(clip = "off") +
+#'   scale_y_reverse(expand = expansion(add = c(0.2, 2))) +
+#'   theme(legend.position = c(0.15, 0.25), legend.background = element_rect())
+#'
+#'
+#' # You can even choose other tree layouts from igraph (e.g. kk)
+#' # and configure multiple styles of custom labels on one plot
+#' set.seed(1) # set seed for reproducibility of ggrepel label positions
+#' actinoOnlyPhylo %>%
+#'   tax_mutate(order = NULL) %>%
+#'   taxatree_label(rank == "family", .label_var = "family_lab") %>%
+#'   taxatree_label(rank == "species", .label_var = "species_lab") %>%
+#'   taxatree_plotkey(
+#'     circular = FALSE, rank != "genus", layout = "kk",
+#'     .draw_label = FALSE, # important not to draw the default labels
+#'     colour = "skyblue", edge_alpha = 0.2
+#'   ) %>%
+#'   taxatree_plot_labels(
+#'     circular = FALSE, label_var = "family_lab", fun = geom_label,
+#'     fontface = "bold.italic", colour = "orange", fill = "black", size = 3
+#'   ) %>%
+#'   taxatree_plot_labels(
+#'     circular = FALSE, label_var = "species_lab", fun = ggrepel::geom_text_repel,
+#'     fontface = "bold.italic", size = 2, force = 20, max.time = 0.1
+#'   )
 taxatree_plotkey <- function(data,
                              ...,
                              size_stat = list(prevalence = prev),
@@ -51,7 +108,7 @@ taxatree_plotkey <- function(data,
 
   # make basic nodes
   if (isTRUE(drop_ranks) && is(data, "psExtra") && !is.null(data@taxatree_stats)) {
-    ranks <- unique(data@taxatree_stats[["rank"]])
+    ranks <- as.character(unique(data@taxatree_stats[["rank"]]))
   } else {
     ranks <- "all"
   }
@@ -70,6 +127,12 @@ taxatree_plotkey <- function(data,
     )
     # join label_data to treeNodes
     treeNodes <- dplyr::left_join(treeNodes, label_data, by = "taxon")
+  }
+  # grab any preceding label info from taxatree stats (not already in nodes df)
+  if (is(data, "psExtra") && !is.null(data@taxatree_stats)) {
+    nonRedundantVars <- setdiff(names(data@taxatree_stats), names(treeNodes))
+    oldLabelData <- data@taxatree_stats[, c("taxon", nonRedundantVars)]
+    treeNodes <- dplyr::left_join(treeNodes, oldLabelData, by = "taxon")
   }
 
   # make graph
@@ -146,6 +209,8 @@ taxatree_plotkey <- function(data,
 #' @param y_nudge
 #' absolute amount by which the initial position of taxon labels is nudged
 #' (relevant only for circular layouts, use nudge_y for other layouts)
+#' @param rotate angle to rotate labels' outer edges away from horizontal
+#' (relevant only for circular layouts, use angle for other layouts)
 #' @param fontface fontface of label text
 #' @param size size of labels
 #' @param colour colour of label outlines and text
@@ -156,6 +221,7 @@ taxatree_plotkey <- function(data,
 #' @param point.padding padding around node points (for label positioning)
 #' @param box.padding padding around labels/text (for label positioning)
 #' @param seed set this for reproducible label positions
+#'
 #' @inheritDotParams ggrepel::geom_text_repel
 #' arrow force force_pull max.time max.iter xlim ylim direction verbose
 #'
@@ -167,6 +233,7 @@ taxatree_plot_labels <- function(p,
                                  label_var = "label",
                                  x_nudge = 0.1,
                                  y_nudge = 0.025,
+                                 rotate = 0,
                                  fontface = "bold",
                                  size = 2.5,
                                  colour = "grey15",
@@ -201,18 +268,20 @@ taxatree_plot_labels <- function(p,
 
   if (isTRUE(circular)) {
     # add circular plot specific nudges and limits
-    args[c("x_nudge", "y_nudge", "xlim", "ylim")] <- list(
-      x_nudge, y_nudge, c(-1.5, 1.5), c(-1.5, 1.5)
+    args[c("x_nudge", "y_nudge", "xlim", "ylim", "rotate")] <- list(
+      x_nudge, y_nudge, c(-1.5, 1.5), c(-1.5, 1.5), rotate
     )
     args[names(dots)] <- dots
+    args <- taxatreePruneLabelArgs(fun = fun, args = args)
 
     for (pos in list(c("x", "y"), "y", NULL, "x")) {
       args[["pos"]] <- pos
       p <- do.call(taxatree_plot_labelsQuadrant, args = args)
-      args[["p"]] <- p
+      args[["p"]] <- p # modified plot used in next iteration of loop
     }
   } else {
     args[names(dots)] <- dots
+    args <- taxatreePruneLabelArgs(fun = fun, args = args)
     p <- do.call(taxatree_plot_labelsNotCircular, args = args)
   }
   return(p)
@@ -226,6 +295,7 @@ taxatree_plot_labelsQuadrant <- function(p,
                                          label_var,
                                          x_nudge,
                                          y_nudge,
+                                         rotate,
                                          ...) {
   neg <- c("x", "y")[!c("x", "y") %in% pos]
 
@@ -246,13 +316,14 @@ taxatree_plot_labelsQuadrant <- function(p,
     data = function(d) {
       dplyr::filter(
         .data = d,
-        dplyr::across(dplyr::all_of(pos), ~ . >= 0),
-        dplyr::across(dplyr::all_of(neg), ~ . < 0),
+        dplyr::if_all(.cols = dplyr::all_of(pos), .fns = ~ . >= 0),
+        dplyr::if_all(.cols = dplyr::all_of(neg), .fns = ~ . < 0),
         .data[[label_var]]
       )
     },
     nudge_x = xsign * x_nudge,
     nudge_y = ysign * y_nudge,
+    angle = ysign * xsign * rotate,
     hjust = hjust
   )
 
@@ -283,4 +354,36 @@ taxatree_plot_labelsNotCircular <- function(p,
 
   p <- p + do.call(what = fun, args = args)
   return(p)
+}
+
+
+#' Helper, removes unwanted arguments from geom_* call args list
+#'
+#' @param fun function name or function itself e.g. geom_label_repel
+#' @param args named list of args
+#'
+#' @return args list possibly with elements removed
+#' @keywords internal
+#' @examples
+#' microViz:::taxatreePruneLabelArgs(fun = "geom_label", args = list(size = 2, seed = 2))
+#' microViz:::taxatreePruneLabelArgs(fun = library, args = list(size = 2, seed = 2))
+taxatreePruneLabelArgs <- function(fun, args) {
+  if (!is.null(args$color)){
+    rlang::warn("'color' argument is ignored, please use 'colour'")
+    args$color <- NULL
+  }
+  # remove ggrepel text or label args from common alternatives from ggplot2
+  # This shouldn't check if NOT ggrepel fun, because a modified version might be used
+  ggRepelArgs <- c(
+    "max.overlaps", "min.segment.length", "segment.size", "segment.color",
+    "segment.colour", "point.padding", "box.padding", "seed", "xlim", "ylim"
+  )
+
+  # label and text separate as there might be further args 1 would want removing
+  if (identical(fun, "geom_label") || identical(fun, ggplot2::geom_label)) {
+    args[ggRepelArgs] <- NULL
+  } else if (identical(fun, "geom_text") || identical(fun, ggplot2::geom_text)) {
+    args[ggRepelArgs] <- NULL
+  }
+  return(args)
 }